feat: move Home into index to add governanceUrl
